Allow to load UI elements from YAML + use properties in Label + Button.
The UI format is a little bit weird, it'll get documented later, but
it's basically YAML with a few conventions that allow you to specify
both content and style at the same place (which is a best practice,
right? right?). Anyway, it works for my usage now, more to come later.